Upgrade tinyrenderer to C++23 and fix all clang-tidy warnings
Replace enable_if with C++20 requires constraints and std::integral
concept, add [[nodiscard]] to const getters, mark single-argument
constructors explicit, rename identifiers to camelBack convention,
add const to local variables, suffix private members with underscore,
refactor nested ternary, and suppress reinterpret_cast in binary I/O.
Also change normalize() to return same type instead of Vec<double, N>.

+++ tinyrenderer/test/VectorTest.cpp
@@ -32,7 +32,7 @@ TEST(Vector, Operators) {
     EXPECT_TRUE(v3 == Vec3f(-3.0f, 6.0f, -3.0f));
 
     // Test normalization
-    Vec3d v4 = v1.normalize();
+    Vec3f v4 = v1.normalize();
     float norm = std::sqrt(1.0f + 4.0f + 9.0f);
     EXPECT_NEAR(v4[0], 1.0f / norm, 1e-6);
     EXPECT_NEAR(v4[1], 2.0f / norm, 1e-6);
